+++ strato/identity-provider/api/.gitignore
@@ -1,0 +1,2 @@
+.stack-work/
+*~
