+++ strato/identity-provider/api/README.md
@@ -1,0 +1,1 @@
+# api
