+++ strato/identity-provider/api/Setup.hs
