+++ strato/identity-provider/client/.gitignore
@@ -1,0 +1,2 @@
+.stack-work/
+*~
