+++ strato/identity-provider/client/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog for `identity-provider`
+# Changelog for `client`
 
 All notable changes to this project will be documented in this file.
