+++ strato/identity-provider/client/README.md
@@ -1,0 +1,1 @@
+# client
