+++ strato/identity-provider/client/package.yaml
@@ -1,0 +1,23 @@
+name:                blockapps-identity-provider-client
+version:             0.1.0.0
+github:              "githubuser/identity-provider-client"
+author:              "Author name here"
+maintainer:          "example@example.com"
+copyright:           "2018 Author name here"
+
+extra-source-files:
+- README.md
+- CHANGELOG.md
+
+
+ghc-options: -Wall -Werror -dynamic
+
+dependencies:
+- base
+- servant-client
+- text
+- blockapps-identity-provider-api
+- strato-model
+
+library:
+  source-dirs: src
