--- strato/identity-provider/package.yaml
@@ -1,92 +1,0 @@
-name:                blockapps-identity-provider
-version:             0.1.0.0
-github:              "githubuser/blockapps-identity-provider"
-author:              "Author name here"
-maintainer:          "example@example.com"
-copyright:           "2023 Author name here"
-
-extra-source-files:
-- README.md
-- CHANGELOG.md
-
-# Metadata used when publishing your package
-# synopsis:            Short description of your package
-# category:            Web
-
-# To avoid duplicated efforts in documentation and dealing with the
-# complications of embedding Haddock markup inside cabal files, it is
-# common to point users to the README.md file.
-description:         Please see the README on GitHub at <https://github.com/githubuser/identity-provider#readme>
-
-dependencies:
-- base
-- common-log
-- slipstream # TODO: switch from slipstream dep to strato-api
-
-ghc-options:
-- -Wall
-- -Wcompat
-- -Widentities
-- -Wincomplete-record-updates
-- -Wincomplete-uni-patterns
-- -Wmissing-export-lists
-- -Wmissing-home-modules
-- -Wpartial-fields
-- -Wredundant-constraints
-- -dynamic
-
-library:
-  source-dirs: src
-  dependencies:
-  - aeson
-  - base64
-  - bloc2api
-  - bloc2client
-  - blockapps-vault-wrapper-client
-  - blockapps-vault-wrapper-api
-  - containers
-  - evm-solidity
-  - hflags
-  - http-client
-  - http-client-tls
-  - http-types
-  - monad-alter
-  - mtl
-  - resourcet
-  - servant-client
-  - servant-server
-  - strato-model
-  - text
-  - unliftio
-  - utf8-string
-  - vault-monad
-  - x509-certs
-
-executables:
-  identity-provider-exe:
-    main:                Main.hs
-    source-dirs:         app
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - aeson
-    - blockapps-identity-provider
-    - bytestring
-    - hflags
-    - http-client
-    - transformers
-    - warp
-    - x509-certs
-
-tests:
-  identity-provider-test:
-    main:                Spec.hs
-    source-dirs:         test
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - blockapps-identity-provider
