--- strato/identity-provider/test/Spec.hs
@@ -1,2 +1,0 @@
-main :: IO ()
-main = putStrLn "Test suite not yet implemented"
